requirements.txt
streamlit
streamlit-webrtc
openai
python-dotenv
elevenlabs
numpy
pydub
